API/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
requests